requirements.txt
# requirements.txt
pyyaml>=6.0
xgboost>=1.7.0
lightgbm>=3.3.0
scikit-learn>=1.2.0
mlflow>=2.0.0
imbalanced-learn>=0.10.0
pandas>=1.5.0
numpy>=1.23.0
pytest>=7.0.0
pytest-cov>=4.0.0
pyspark>=3.3.0